+++ packages/twenty-server/src/modules/connected-account/auto-companies-and-contacts-creation/create-company/create-company.service.ts
@@ -5,7 +5,7 @@ import { v4 } from 'uuid';
 import axios, { AxiosInstance } from 'axios';
 
 import { CompanyRepository } from 'src/modules/company/repositories/company.repository';
-import { getCompanyNameFromDomainName } from 'src/modules/messaging/utils/get-company-name-from-domain-name.util';
+import { getCompanyNameFromDomainName } from 'src/modules/participant/utils/get-company-name-from-domain-name.util';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
 import { CompanyObjectMetadata } from 'src/modules/company/standard-objects/company.object-metadata';
 @Injectable()
