+++ packages/twenty-server/src/modules/participant/utils/get-company-name-from-domain-name.util.ts
