+++ packages/twenty-server/src/modules/participant/utils/get-domain-name-from-handle.util.ts
