+++ packages/twenty-server/src/modules/participant/utils/get-first-name-and-last-name-from-handle-and-display-name.util.ts
