+++ packages/twenty-server/src/modules/participant/utils/is-person-email.util.ts
